调整命令类

+++ tutelary-message/src/main/java/com/tutelary/message/command/GarbageCollector.java
@@ -10,7 +10,7 @@ import java.io.Serializable;
 public class GarbageCollector implements Serializable {
 
     private String name;
-    private Integer collectionCount;
-    private Integer collectionTime;
+    private long collectionCount;
+    private long collectionTime;
 
 }
